Refactor chart line module
